backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Bcrypt
Flask-CORS
psycopg2-binary
python-dotenv
Werkzeug
